refactor, style: Divison, Gender, LeaderRole enum class member 도메인 패키지에서 global 도메인 패키지로 이동 및 전체 프로젝트 optimize code 적용
