+++ src/main/java/com/windstorm/management/controller/member/response/MemberResponse.java
@@ -2,9 +2,9 @@ package com.windstorm.management.controller.member.response;
 
 import java.time.LocalDate;
 
-import com.windstorm.management.domain.member.Division;
-import com.windstorm.management.domain.member.Gender;
-import com.windstorm.management.domain.member.LeaderRole;
+import com.windstorm.management.domain.global.Division;
+import com.windstorm.management.domain.global.Gender;
+import com.windstorm.management.domain.global.LeaderRole;
 import com.windstorm.management.domain.member.Member;
 
 import lombok.Builder;
