+++ src/main/java/com/windstorm/management/security/jwt/JwtProvider.java
@@ -11,11 +11,9 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
-import com.windstorm.management.implement.member.MemberReader;
 import com.windstorm.management.security.CustomUserDetailsService;
 
 import io.jsonwebtoken.Claims;
